Refactor SongForm to be more single-responsibility
