+++ backend/app/controllers/api/v1/songs_controller.rb
@@ -4,7 +4,7 @@ class Api::V1::SongsController < ApplicationController
         song = Song.new(song_params)
         song.user_id = @user.id
         if song.save
-            render json: SongSerializer.new(song), status: :created
+            render json: song, status: :created
         else
             render json: { error: 'Failed to create song', messages: song.errors.full_messages }, status: :bad_request
         end
