+++ frontend/src/components/songs/NewSong.js
@@ -1,6 +1,7 @@
 import React, { Component } from 'react'
 import SearchContainer from '../search/SearchContainer'
-import SongForm from './SongForm'
+import SongFormContainer from './SongFormContainer'
+import { getAuthTokenHeader, backendURL } from '../../utils'
 
 
 class NewSong extends Component {
@@ -35,17 +36,47 @@ class NewSong extends Component {
     })
   }
 
+  // Form submit action that is specific to creating a new song (must pass spotify data and form data to backend)
+  handleSubmit = (event, formData) => {
+    event.preventDefault()
+    
+    // Create new object to aggregate spotify and user-entered form data to be sent to the server
+    const songData = {}
+
+    // Aggregate the spotify and form data, renaming `sections` to `sections_attributes` which is what the server expects
+    delete Object.assign(songData, this.state.spotifyData, formData, { sections_attributes: formData.sections }).sections
+
+    // Each section needs to have a `display_order` attribute at time of submit. Use the order of the array.
+    songData.sections_attributes = songData.sections_attributes.map((section, index) => ({...section, display_order: index + 1}))
+
+    const req = {
+      method: 'POST',
+      headers:{
+        ...getAuthTokenHeader(),
+        'Content-Type': 'application/json'
+      },
+      body: JSON.stringify({
+        song: songData
+      })
+    }
+
+    // TODO: Standardize with rest of project
+    fetch(`${backendURL}/songs`, req)
+    .then(resp => resp.json())
+    .then(song => console.log('Data returned:', song))
+  }
+
   render() {
     return (
-      <>
+      <div>
         <h3>New Song</h3>
         {
           this.state.displaySearch ?
           <SearchContainer handleSpotifyData={this.handleSpotifyData}/>
             :
-          <SongForm spotifyData={this.state.spotifyData}/>
+          <SongFormContainer spotifyData={this.state.spotifyData} handleSubmit={this.handleSubmit}/>
         }
-      </>
+      </div>
     )
   }
 }
